requirements.txt
pex~=2.46.2
httpx~=0.28.1
transformers~=4.53.2
huggingface_hub~=0.33.4
filelock~=3.18.0
requests~=2.32.4
qdrant-client~=1.15.0
pydantic~=2.11.7
coloredlogs~=15.0.1
langchain-text-splitters~=0.3.8
spacy~=3.8.7
diskcache~=5.6.3
keybert~=0.9.0
scikit-learn~=1.7.1
validators~=0.35.0
tiktoken~=0.9.0
psutil~=7.0.0
GitPython~=3.1.44
toml~=0.10.2
ollama~=0.5.1
numpy~=2.3.1
networkx~=3.5